requirements.txt
flask
xhtml2pdf
python-dotenv
markupsafe
openai
gunicorn
sqlalchemy
psycopg2-binary
werkzeug

Flask-Mail
pytz
